Rename directory CheckStatusResponses to check-status-responses
