+++ __tests__/components/check-status-responses/CheckStatusFileBeingProcessed.test.tsx
@@ -3,17 +3,17 @@ import { render, screen } from '@testing-library/react'
 
 import { axe, toHaveNoViolations } from 'jest-axe'
 
-import CheckStatusNoRecord from '../../components/CheckStatusResponses/CheckStatusNoRecord'
+import CheckStatusFileBeingProcessed from '../../../components/check-status-responses/CheckStatusFileBeingProcessed'
 
 expect.extend(toHaveNoViolations)
 
-describe('CheckStatusNoRecord', () => {
-  const sut = <CheckStatusNoRecord />
+describe('CheckStatusFileBeingProcessed', () => {
+  const sut = <CheckStatusFileBeingProcessed />
 
   //TODO: add test for when phone number is visible and when it isn't
   it('renders', () => {
     render(sut)
-    expect(screen.getByTestId('no-record')).toBeInTheDocument()
+    expect(screen.getByTestId('being-processed')).toBeInTheDocument()
   })
 
   it('meets a11y', async () => {
