+++ __tests__/components/check-status-responses/CheckStatusNotAcceptable.test.tsx
@@ -3,17 +3,16 @@ import { render, screen } from '@testing-library/react'
 
 import { axe, toHaveNoViolations } from 'jest-axe'
 
-import CheckStatusFileBeingProcessed from '../../components/CheckStatusResponses/CheckStatusFileBeingProcessed'
+import CheckStatusNotAcceptable from '../../../components/check-status-responses/CheckStatusNotAcceptable'
 
 expect.extend(toHaveNoViolations)
 
-describe('CheckStatusFileBeingProcessed', () => {
-  const sut = <CheckStatusFileBeingProcessed />
+describe('CheckStatusNotAcceptable', () => {
+  const sut = <CheckStatusNotAcceptable />
 
-  //TODO: add test for when phone number is visible and when it isn't
   it('renders', () => {
     render(sut)
-    expect(screen.getByTestId('being-processed')).toBeInTheDocument()
+    expect(screen.getByTestId('not-acceptable')).toBeInTheDocument()
   })
 
   it('meets a11y', async () => {
