+++ __tests__/components/check-status-responses/CheckStatusShippingCanadaPost.test.tsx
@@ -3,16 +3,17 @@ import { render, screen } from '@testing-library/react'
 
 import { axe, toHaveNoViolations } from 'jest-axe'
 
-import CheckStatusNotAcceptable from '../../components/CheckStatusResponses/CheckStatusNotAcceptable'
+import CheckStatusShippingFedex from '../../../components/check-status-responses/CheckStatusShippingFedex'
 
 expect.extend(toHaveNoViolations)
 
-describe('CheckStatusNotAcceptable', () => {
-  const sut = <CheckStatusNotAcceptable />
+describe('CheckStatusShippingFedex', () => {
+  const sut = <CheckStatusShippingFedex />
 
+  //TODO: add test for when phone number is visible and when it isn't
   it('renders', () => {
     render(sut)
-    expect(screen.getByTestId('not-acceptable')).toBeInTheDocument()
+    expect(screen.getByTestId('shipped-fedex')).toBeInTheDocument()
   })
 
   it('meets a11y', async () => {
